+++ CHANGELOG.md
@@ -1,9 +1,12 @@
+# Obsidian to latex changelog
+
 ### New Features
 1. Support nested lists indented lists
 2. Support numbered lists
 3. Support numbered lists starting from numbers other than 1
 4. Convert markdown links to clickable links
 5. Barely support inline code snippets using `back ticks`
+6. Accept latex template from user
 
 ### Fixes